// src/app/utils/seedSuperAdmin.ts
import bcryptjs from 'bcryptjs'
import { envVers } from '../config/env'
import { IAuthProvider, IUser, Role } from '../modules/user/user.interface'
import { User } from '../modules/user/user.model'

export const seedSuperAdmin = async () => {
  try {
    const isSuperAdminExist = await User.findOne({
      email: envVers.SUPER_ADMIN_EMAIL,
    })

    if (isSuperAdminExist) {
      console.log('Super Admin Already Exists!')
      return
    }

    console.log('Trying to create Super Admin...')

    const hashedPassword = await bcryptjs.hash(
      envVers.SUPER_ADMIN_PASSWORD,
      Number(envVers.BCRYPT_SALT_ROUND)
    )

    const authProvider: IAuthProvider = {
      provider: 'credentials',
      providerId: envVers.SUPER_ADMIN_EMAIL,
    }

    const payload: IUser = {
      name: 'Super admin',
      role: Role.SUPER_ADMIN,
      email: envVers.SUPER_ADMIN_EMAIL,
      password: hashedPassword,
      isVerified: true,
      auths: [authProvider],
    }

    const superadmin = await User.create(payload)
    console.log('Super Admin Created Successfuly! \n')
    console.log(superadmin)
  } catch (error) {
    console.log(error)
  }
}
